requirements.txt
Flask==1.1.2
flask-wtf
wtforms
email_validator
flask-paginate
Pillow
boto3
requests